fixes: pkvm: x86: Change pkvm_page order to u8

+++ arch/x86/kvm/pkvm/memory.h
@@ -12,7 +12,7 @@
 
 struct pkvm_page {
 	unsigned short refcount;
-	unsigned short order;
+	u8 order;
 };
 
 extern u64 __pkvm_vmemmap;
